Stop truncating econ_order's score and steps to int
The C prints both with %ld (score/steps are long). The Java ported
%d with an (int) cast, silently truncating to 32 bits for large
values instead of translating a C long %ld to a Java %d applied to
the long directly (Java's %d formats a long natively — no narrowing
needed).

Honest exception: pure formatting fix, no behavior difference on any
captured oracle case (all fit in 32 bits); verified by the full build
gate (spotlessApply + check) rather than a new unit test.

+++ demos/src/main/java/com/robsartin/jsgb/demo/EconOrder.java
@@ -166,7 +166,7 @@ public final class EconOrder {
       }
 
       if (Gb.verbose != 0) {
-        out.print(String.format(Locale.ROOT, "%8d after step %d\n", (int) score, (int) steps));
+        out.print(String.format(Locale.ROOT, "%8d after step %d\n", score, steps));
       } else if (steps % 1000 == 0 && steps > 0) {
         out.print(".");
       }
@@ -210,8 +210,8 @@ public final class EconOrder {
             Locale.ROOT,
             "\n%s is %d, found after %d step%s.\n",
             bestScore == INF ? "Local minimum feed-forward" : "Another local minimum",
-            (int) score,
-            (int) steps,
+            score,
+            steps,
             steps == 1 ? "" : "s"));
     if (Gb.verbose != 0 || score < bestScore) {
       out.print("The corresponding economic order is:\n");
